[vfs] E: Split Terminal Domain Types
Move LineDiscipline and its private segment helper into a dedicated
discipline module, and place console read outcomes and terminal signals
in separate modules. Preserve the existing line_discipline exports and
terminal behavior.
